Formatting and a better comment

+++ src/main.cpp
@@ -81,7 +81,7 @@ void print_logo() {
 
 // Master loop
 int main() {
-    long long loopc = 0; // Loop counter for heartbeat
+    uint32_t loopc = 0; // Loop counter for heartbeat
 
     Serial.begin(112500); // the serial monitor is actually always active (for debug use Serial.println & tycmd)
     debug.begin(SerialUSB1);
@@ -112,7 +112,6 @@ int main() {
 
     //generate controller outputs based on governed references and estimated state
     controller_manager.init(&can, config);
-
 
     //set reference limits in the reference governor
     governor.set_reference_limits(config->set_reference_limits);
@@ -228,6 +227,7 @@ int main() {
                 hive_toggle = false;
             }
         }
+
         // when in teensy control mode reset hive toggle
         if (dr16.get_l_switch() == 3) {
             if (!hive_toggle) {
@@ -236,7 +236,6 @@ int main() {
             }
             hive_toggle = true;
         }
-
 
         // read sensors
         estimator_manager.read_sensors();
@@ -247,10 +246,8 @@ int main() {
             memcpy(temp_state, hive_state_offset, sizeof(hive_state_offset));
         }
 
-
         // step estimates and construct estimated state
         estimator_manager.step(temp_state, temp_micro_state, incoming->get_hive_override_request());
-
 
         // if first loop set target state to estimated state
         if (count_one == 0) {
@@ -264,12 +261,8 @@ int main() {
         governor.step_reference(target_state, config->governor_types);
         governor.get_reference(temp_reference);
 
-        // Serial.printf("yaw ref: %f, pitch ref: %f\n", temp_reference[3][0], temp_reference[4][0]);
-
-
         // generate motor outputs from controls
         controller_manager.step(temp_reference, temp_state, temp_micro_state);
-
 
         // construct sensor data packet
         SensorData sensor_data;
@@ -302,7 +295,9 @@ int main() {
 	    float dt = stall_timer.delta();
         if (dt > 0.002) { 
 	    	Serial.printf("Slow loop with dt: %f\n", dt);
+            // zero the can bus just in case
 	    	can.zero();
+            // mark this as a slow loop to trigger safety mode
 	    	is_slow_loop = true;
 	    }
         
@@ -322,7 +317,6 @@ int main() {
 
         // Keep the loop running at the desired rate
         loop_timer.delay_micros((int)(1E6 / (float)(LOOP_FREQ)));
-        
     }
     
     return 0;
